webscraper-fastapi-dev-uat/requirements.txt
# Core API Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
python-multipart>=0.0.12
websockets>=13.0

# Environment and Configuration
python-dotenv>=1.0.1

# Template Engine for GUI
jinja2>=3.1.4

# AI and Web Scraping
google-generativeai>=0.8.0
langchain-google-genai>=0.0.15

# Web Scraping and Browser Automation
playwright>=1.51.0
aiohttp>=3.11.0
beautifulsoup4>=4.12.3
lxml>=5.3.0

# Data Processing and Utilities
dataclasses-json>=0.6.7
requests>=2.32.0
urllib3>=2.2.0

# HTTP Client for async operations
httpx>=0.28.0

# Date and time utilities
python-dateutil>=2.9.0

# JSON handling
orjson>=3.10.0

# Async utilities
asyncio-throttle>=1.0.2

# File handling
aiofiles>=24.1.0

# Regular expressions (already in stdlib but explicit for clarity)
regex>=2024.11.0

# Logging and monitoring
structlog>=24.1.0

# Additional utilities for production
gunicorn>=23.0.0  # For production WSGI server
psutil>=6.1.0     # System monitoring
click>=8.1.7      # CLI utilities

# Development and testing dependencies (optional)
pytest>=8.0.0
pytest-asyncio>=0.24.0
black>=24.0.0
isort>=5.13.0
mypy>=1.8.0

# Security
cryptography>=43.0.0

# Optional: Database drivers (uncomment if needed)
# sqlalchemy>=2.0.0
# psycopg2-binary>=2.9.0  # PostgreSQL
# redis>=5.2.0            # Redis

pillow==10.0.1
cachetools==5.3.2
tenacity==8.2.3